pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hydraflow"
version = "0.8.0"
description = "Hydraflow integrates Hydra and MLflow to manage and track machine learning experiments."
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "daizutabi", email = "daizutabi@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = ["hydra-core>=1.3", "mlflow>=2.15", "omegaconf", "rich", "typer"]

[project.urls]
Documentation = "https://daizutabi.github.io/hydraflow/"
Source = "https://github.com/daizutabi/hydraflow"
Issues = "https://github.com/daizutabi/hydraflow/issues"

[project.scripts]
hydraflow = "hydraflow.cli:app"

[dependency-groups]
dev = [
  "pytest-clarity",
  "pytest-cov",
  "pytest-order",
  "pytest-randomly",
  "pytest-xdist",
]
docs = ["markdown-exec[ansi]", "mkapi", "mkdocs-material"]

[tool.pytest.ini_options]
addopts = [
  "--cov=hydraflow",
  "--cov-report=lcov:lcov.info",
  "--dist=loadgroup",
  "-n8",
]
filterwarnings = [
  "ignore:Support for class-based `config` is deprecated",
  "ignore:Pydantic V1 style",
]

[tool.coverage.report]
exclude_lines = ["no cov", "raise NotImplementedError", "if TYPE_CHECKING:"]
skip_covered = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["ALL"]
unfixable = ["F401"]
ignore = [
  "A005",
  "ANN003",
  "ANN401",
  "B904",
  "D104",
  "D105",
  "D107",
  "D203",
  "D213",
  "EM101",
  "FBT001",
  "FBT002",
  "PGH003",
  "PLR0911",
  "PLR0913",
  "PLR1704",
  "PLR2004",
  "SIM102",
  "SIM108",
  "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["A001", "ANN", "ARG", "D", "FBT", "PD", "PLR", "PT", "S", "SLF"]
"apps/*.py" = ["D", "G", "INP"]
"src/hydraflow/main.py" = ["ANN201", "D401"]
"src/hydraflow/cli.py" = ["ANN", "D"]
